feat(styling): added intermediate config / styling for error reporting
feat #1504
