+++ src/styles/config.ts
@@ -1,0 +1,9 @@
+enum StyleVersion {
+  Intermediate,
+  Final,
+}
+
+const styleVersion = StyleVersion.Intermediate
+
+// @ts-ignore
+export const isFinalStyle = () => styleVersion === StyleVersion.Final
